pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "survey-analysis-agent"
version = "0.1.0"
description = "An Intelligent Agent for Probabilistic and Context-Aware Analysis of Free-Text Survey Responses"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.2",
    "fastapi>=0.136.0",
    "hmmlearn>=0.3.3",
    "jinja2>=3.1.6",
    "nltk>=3.9.3",
    "pandas>=2.0.0,<3.0.0",
    "pgmpy>=1.0.0",
    "pydantic>=2.13.2",
    "pydantic-ai-slim[bedrock,openai]>=1.84.1",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.44.0",
]

[project.scripts]
download-data = "scripts.download_data:main"
preprocess-data = "scripts.preprocess_data:main"
tokenize-sentences = "scripts.tokenize_sentences:main"
create-sample = "scripts.create_sample:main"
label-data = "scripts.label_training_data:main"
validate-data = "scripts.validate_data:main"
prepare-data = "scripts.prepare_all_data:main"

[tool.setuptools]
packages = ["scripts", "src"]

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "streamlit>=1.55.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
